+++ src/compliance_tests/TestIso_7_6_13.cpp
@@ -75,7 +75,7 @@ class TestIso_7_6_13 : public test::TestBase
         void ConfigureTest()
         {
             FillTestVariants(VariantMatchType::CommonAndFd);
-            for (int i = 0; i < 2; i++)
+            for (size_t i = 0; i < 2; i++)
             {
                 AddElemTest(TestVariant::Common, ElemTest(i + 1, FrameKind::Can20));
                 AddElemTest(TestVariant::CanFdEna, ElemTest(i + 1, FrameKind::CanFd));
@@ -111,15 +111,15 @@ class TestIso_7_6_13 : public test::TestBase
             drv_bit_frm->InsertOvrlFrm(0, BitKind::Interm);
 
             /*  Force n-th bit of Overload Delimiter to Dominant */
-            int bit_to_corrupt;
+            size_t bit_to_corrupt;
             if (elem_test.index_ == 1)
                 bit_to_corrupt = 2;
             else
                 bit_to_corrupt = 7;
 
-            TestMessage("Forcing Overload delimiter bit %d to recessive", bit_to_corrupt);
+            TestMessage("Forcing Overload delimiter bit %zu to recessive", bit_to_corrupt);
             Bit *bit = drv_bit_frm->GetBitOf(bit_to_corrupt - 1, BitKind::OvrlDelim);
-            int bit_index = drv_bit_frm->GetBitIndex(bit);
+            size_t bit_index = drv_bit_frm->GetBitIndex(bit);
             bit->val_ = BitVal::Dominant;
 
             /* Insert Error flag from one bit further, both driver and monitor! */
